fix(go-rewrite): dedupe round-3 helper redeclarations
Wave 2 round-3 PRs (GetEdges{From,To}, Get{Node,Nodes}, Search/QueryNodes,
sharing/ACL ports, ...) each independently added the same helpers
(edgeToProto, role*, storeNodeToProto, authActorToACLActor,
storeVisibilityAdapter, newIdempotencyKey). When all merged onto main,
`go vet ./internal/api/...` failed with redeclarations.

Consolidate into helpers.go and helpers_external_test.go (same pattern
as fix/go-api-duplicate-helpers in #460).

Refs #407
